+++ packages/fx-core/tests/component/coordinator/coordinator.deploy.test.ts
@@ -23,7 +23,7 @@ import { metadataUtil } from "../../../src/component/utils/metadataUtil";
 import { pathUtils } from "../../../src/component/utils/pathUtils";
 import { settingsUtil } from "../../../src/component/utils/settingsUtil";
 import { FxCore } from "../../../src/core/FxCore";
-import { setTools } from "../../../src/core/globalVars";
+import { setTools } from "../../../src/common/globalVars";
 import * as v3MigrationUtils from "../../../src/core/middleware/utils/v3MigrationUtils";
 import { UserCancelError } from "../../../src/error";
 import { MockTools } from "../../core/utils";
