+++ packages/fx-core/tests/component/driver/arm/deploy.test.ts
@@ -4,7 +4,7 @@
 import { assert } from "chai";
 import "mocha";
 import { createSandbox } from "sinon";
-import { setTools } from "../../../../src/core/globalVars";
+import { setTools } from "../../../../src/common/globalVars";
 import {
   MockAzureAccountProvider,
   MockLogProvider,
