+++ packages/fx-core/tests/component/driver/oauth/update.test.ts
@@ -12,7 +12,7 @@ import {
   MockedM365Provider,
   MockedUserInteraction,
 } from "../../../plugins/solution/util";
-import { setTools } from "../../../../src/core/globalVars";
+import { setTools } from "../../../../src/common/globalVars";
 import { AppStudioClient } from "../../../../src/component/driver/teamsApp/clients/appStudioClient";
 import { UpdateOauthDriver } from "../../../../src/component/driver/oauth/update";
 import {
