+++ packages/fx-core/tests/component/middleware/middleware.test.ts
@@ -2,7 +2,7 @@ import "mocha";
 import * as chai from "chai";
 import chaiAsPromised from "chai-as-promised";
 import { MockTools } from "../../core/utils";
-import { setTools } from "../../../src/core/globalVars";
+import { setTools } from "../../../src/common/globalVars";
 import { MockDriver } from "./helper";
 import sinon from "sinon";
 import { TelemetryConstants } from "../../../src/component/constants";
