+++ packages/fx-core/tests/core/middleware/VideoFilterAppBlockerMW.test.ts
@@ -8,7 +8,7 @@ import "mocha";
 import mockFs from "mock-fs";
 import * as path from "path";
 import { VideoFilterAppRemoteNotSupportedError } from "../../../src/core/error";
-import { setTools } from "../../../src/core/globalVars";
+import { setTools } from "../../../src/common/globalVars";
 import { VideoFilterAppBlockerMW } from "../../../src/core/middleware/videoFilterAppBlocker";
 import { CoreHookContext } from "../../../src/core/types";
 import { MockTools } from "../utils";
