+++ packages/fx-core/tests/ui/qm.visitor.test.ts
@@ -44,7 +44,7 @@ import { assert } from "chai";
 import "mocha";
 import mockedEnv, { RestoreFn } from "mocked-env";
 import sinon from "sinon";
-import { setTools } from "../../src/core/globalVars";
+import { setTools } from "../../src/common/globalVars";
 import {
   EmptyOptionError,
   InputValidationError,
